bug fixes and improvements

+++ app/src/main/java/com/vanced/manager/core/downloader/MicrogDownloadService.kt
@@ -45,7 +45,7 @@ class MicrogDownloadService: Service() {
                     .build()
                     .setOnProgressListener { progress ->
                         val mProgress = progress.currentBytes * 100 / progress.totalBytes
-                        localBroadcastManager.sendBroadcast(Intent(HomeFragment.MICROG_DOWNLOADING).putExtra("progress", mProgress).putExtra("file", getFileNameFromUrl(url)))
+                        localBroadcastManager.sendBroadcast(Intent(HomeFragment.MICROG_DOWNLOADING).putExtra("progress", mProgress.toInt()).putExtra("file", getFileNameFromUrl(url)))
                     }
                     .start(object : OnDownloadListener {
                         override fun onDownloadComplete() {
